Move grid globals to grid tab within addons
