+++ polaris-react/.storybook/GridPanel/index.ts
@@ -1,1 +1,2 @@
 export * from './GridOverlay';
+export * from './GridPanel';
