+++ polaris-react/.storybook/main.js
@@ -20,7 +20,6 @@ module.exports = {
     {name: '@storybook/addon-essentials', options: {docs: enableDocs}},
     '@storybook/addon-a11y',
     'storybook-addon-performance/register',
-    '@luigiminardim/storybook-addon-globals-controls',
   ],
   webpackFinal: (config) => {
     const isProduction = config.mode === 'production';
