+++ polaris-react/package.json
@@ -71,7 +71,8 @@
     "react-dom": "^16.14.0 || ^17.0.0"
   },
   "devDependencies": {
-    "@luigiminardim/storybook-addon-globals-controls": "^0.1.3",
+    "@babel/core": "^7.15.0",
+    "@babel/node": "^7.14.9",
     "@rollup/plugin-image": "^2.0.5",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-replace": "^2.3.3",
